Class Function Split
