+++ src/pages/matrix/MatrixDemo.vue
@@ -1,9 +1,10 @@
 <template>
-    <matrix-login v-if="!$matrix.isLoggedIn()"></matrix-login>
-    <q-page v-if="$matrix.isLoggedIn()">
+    <!--<matrix-login v-if="!$matrix.isLoggedIn()"></matrix-login>-->
+    <!--<q-page v-if="$matrix.isLoggedIn()">-->
+    <q-page>
       <div>
         Matrix Demo Seite funktioniert!
-        <q-btn color="purple" label="Test" @click="() => testfunction()"/>
+        <q-btn color="purple" label="Login" @click="() => login()"/>
         <q-btn color="purple" label="Daten" @click="() => checkStorage()"/>
       <q-btn
           color="black"
@@ -20,18 +21,18 @@
 
 <script lang="ts">
 import {defineComponent} from 'vue';
-import MatrixLogin from '../../components/MatrixLogin.vue'
+//import MatrixLogin from '../../components/MatrixLogin.vue'
 
 
 
 export default defineComponent({
   name: 'MatrixDemo',
   components: {
-      'matrix-login': MatrixLogin
+      //'matrix-login': MatrixLogin
   },
   methods: {
-    testfunction(){
-        this.$matrix.consoleMtxClient();
+    login(){
+      this.$matrix.login();
     },
     checkStorage(){
       this.$matrix.checkStorage();
